fix(agents): bind fetch for Events oRPC client in Workers
OpenAPILink must not receive an unbound `fetch` reference (illegal invocation
in Cloudflare Workers). Handle POST /api/events-forwarded inside the /api/$
catch-all before the oRPC handler so webhooks are not 404’d.

Made-with: Cursor
